[#3] refactor: InfoContainor 컴포넌트 분리
날짜와 마지막 동기화 시간을 보여주는 기능을 컴포넌트로 분리
MainPage에서 확인 가능하도록 추상화
