+++ src/pages/MainPage.tsx
@@ -1,24 +1,30 @@
 import { useMemo } from 'react';
 import useOrderList from 'hooks/useOrderList';
 import useColums from 'hooks/useColums';
-import LayoutWrapper from 'components/common/LayoutWrapper';
+import LayoutWrapper from 'components/layout/LayoutWrapper';
 import CustomSkeleton from 'components/common/CustomSkeleton';
-import MainTable from 'components/MainTable';
+import MainTable from 'components/table/MainTable';
+import { getOrderListDataApi } from 'utils/api';
+import { useQuery } from '@tanstack/react-query';
+import InfoContainor from 'components/tools/InfoContainor';
 
 const MainPage = () => {
-  const { data: orders, isLoading, dataUpdatedAt } = useOrderList();
+  const { data, isLoading } = useQuery({
+    queryKey: ['orderList'],
+    queryFn: getOrderListDataApi,
+    // refetchInterval: 5000,
+  });
   const columns = useColums();
-  const data = useMemo(() => orders, [orders]);
 
   if (isLoading || !data) return <CustomSkeleton />;
 
   return (
     <LayoutWrapper>
+      <InfoContainor />
       <MainTable
         {...{
           data,
           columns,
-          dataUpdatedAt,
         }}
       />
     </LayoutWrapper>
